refactor(library): removed file for client options (#16)
